template config with languages
#9
It gets the device language and sets the var 'navigator.app_lang' to an object that has all the language vars accordingly.

Each page will use `navigator.app_lang` and its vars to set everything in that language, like `navigator.app_lang.picture` will have the word 'תמונה' if the device is in hebrew.

All the language configurations are set in tconfig.json and must have all the vars that can be used.